build: make LunEx into a Maven module
