+++ .gitignore
@@ -1,4 +1,4 @@
-*.class
+target/
 
 # Mobile Tools for Java (J2ME)
 .mtj.tmp/
