+++ LunEx/pom.xml
@@ -3,14 +3,13 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.agileinstitute</groupId>
-    <artifactId>labs-java-junit</artifactId>
-    <version>1.0-SNAPSHOT</version>
-    <packaging>pom</packaging>
+    <parent>
+        <groupId>com.agileinstitute</groupId>
+        <artifactId>labs-java-junit</artifactId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
 
-    <modules>
-        <module>LunEx</module>
-    </modules>
+    <artifactId>LunEx</artifactId>
 
     <properties>
         <maven.compiler.release>21</maven.compiler.release>
