--- LunEx/src/ITC/SecurityExchangeTransmissionInterface.java
@@ -1,5 +1,0 @@
-package ITC;
-
-public interface SecurityExchangeTransmissionInterface {
-	int currentPrice(String symbol);
-}
