--- LunEx/src/lunEx/LunExServiceUnavailableException.java
@@ -1,7 +1,0 @@
-package lunEx;
-@SuppressWarnings("serial")
-public class LunExServiceUnavailableException extends RuntimeException {
-	public String getMessage() {
-		return "Sorry, sunspot activity today...please try again later";
-	}
-}
