+++ LunEx/src/main/java/ITC/SecurityExchangeTransmissionInterface.java
@@ -1,0 +1,5 @@
+package ITC;
+
+public interface SecurityExchangeTransmissionInterface {
+	int currentPrice(String symbol);
+}
