+++ LunEx/src/main/java/lunEx/LunExServiceUnavailableException.java
@@ -1,0 +1,7 @@
+package lunEx;
+@SuppressWarnings("serial")
+public class LunExServiceUnavailableException extends RuntimeException {
+	public String getMessage() {
+		return "Sorry, sunspot activity today...please try again later";
+	}
+}
